refactor(DateRangePicker): rename DatePicker to DateRangePicker
