--- src/components/Filters/DatePicker/index.ts
@@ -1,1 +1,0 @@
-export { default as DatePicker } from './DatePicker';
