+++ src/components/Filters/DateRangePicker/index.ts
@@ -1,0 +1,1 @@
+export { default as DateRangePicker } from './DateRangePicker';
