+++ src/components/Filters/mocks/options.ts
@@ -1,5 +1,5 @@
 import { Input } from '../Input';
-import { DatePicker } from '../DatePicker';
+import { DateRangePicker } from '../DateRangePicker';
 import { DataPoint, Filter } from '../Filters.types';
 import { Operators } from '../Filters.enums';
 
@@ -37,7 +37,7 @@ export const dataPointsMock: DataPoint[] = [
         isDefault: true,
       },
       { input: Input, value: 'strawberry', label: 'Strawberry' },
-      { input: DatePicker, value: 'vanilla', label: 'Vanilla' },
+      { input: DateRangePicker, value: 'vanilla', label: 'Vanilla' },
     ],
     value: 'smurf',
     label: 'Smurf',
